requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
httpx==0.25.2
openai==1.30.5
